+++ code/ext/ext.lib.monaco.crdt/src/index.pkg.mts
@@ -17,6 +17,7 @@ export const Pkg: ModuleDef = {
     'react-dom': '18.2.0',
     'ext.lib.automerge': '0.0.0',
     'ext.lib.monaco': '0.0.0',
+    'sys.data.json': '0.0.0',
     'sys.util': '0.0.0',
     'sys.ui.react.common': '0.0.0',
   },
